+++ src/promoted/code1_0029.c
@@ -43,7 +43,7 @@ s32 func_004598e0(s16 arg0);
 s32 func_00104ce0(s16 arg0);
 void func_0029cf50(s32 arg0);
 
-s32 func_00104dc0(s16 arg0);
+u16 func_00104dc0(s32 arg0);
 
 s32 func_00104d50(s16 arg0);
 extern void func_001056e0(s16 arg0, s16 arg1);
@@ -58,7 +58,7 @@ extern u8 D_0063CFA0[];
 extern s32 func_00452380(void *path);
 extern s32 func_00106330(s32 arg0);
 extern void func_001238c0(s32 arg0);
-extern s32 func_00104e30(s16 arg0);
+extern u32 func_00104e30(s32 arg0);
 extern void func_00105730(s16 arg0, s16 arg1);
 static inline u8 *code29AddOff(s32 offset, u8 *base) {
     return (u8 *)((u32)offset + (u32)base);
@@ -835,7 +835,7 @@ s32 func_00299c60(void) {
 
 // FUN_00299CB0
 s32 func_00299cb0(void) {
-    func_0029cf50(func_00104dc0(func_0029cc00(0)) & 0xFFFF);
+    func_0029cf50(func_00104dc0((s16)func_0029cc00(0)) & 0xFFFF);
     return 1;
 }
